Restructure directory layout
