+++ app/(protected)/mine/page.tsx
@@ -1,0 +1,5 @@
+import { RedirectType, redirect } from 'next/navigation';
+
+export default async function Page() {
+  redirect('/mine/collections', RedirectType.replace);
+}
