+++ app/(protected)/settings/layout.tsx
@@ -1,27 +1,15 @@
 import { PropsWithChildren } from 'react';
-import { headers } from 'next/headers';
 import { SettingsNavRoutes } from '@protected/settings/SettingsNavRoutes';
+import TopSubNav from '@components/navs/TopSubNav';
 
 export default async function SettingsLayout({ children }: PropsWithChildren) {
-  const pathname = headers().get('x-next-pathname') as string;
-
   return (
     <div className="" data-testid='settings-layout'>
       <h1 className="sr-only">Settings</h1>
 
       <header className="border-b border-white/5">
         {/* Secondary navigation */}
-        <nav className="flex overflow-x-auto">
-          <ul role="list" className="flex min-w-full flex-none gap-x-6 text-sm font-semibold leading-6 text-gray-400">
-            {SettingsNavRoutes.map((item) => (
-              <li key={item.name}>
-                <a href={item.href} className={item.href === pathname ? 'text-indigo-400' : ''}>
-                  {item.name}
-                </a>
-              </li>
-            ))}
-          </ul>
-        </nav>
+        <TopSubNav navItems={SettingsNavRoutes} />
       </header>
   
       <div className="">
