+++ app/(protected)/subscriptions/collections/page.tsx
@@ -1,8 +1,8 @@
 import { getAuthSession } from '@app/auth/session';
-import FunctionsList from './FunctionList';
+import CollectionList from './CollectionList';
 import { cn } from '@lib/utils';
 import { notFound } from 'next/navigation';
-import type { FetchFunctionSubscriptionListPaginationType } from '@protected/mine/subscriptions/functions/types';
+import type { FetchCollectionSubscriptionListPaginationType } from '@protected/subscriptions/collections/types';
 import { BakaPagination } from '@components/BakaPagination';
 import clamp from 'lodash/clamp';
 import db from '@lib/db';
@@ -19,53 +19,61 @@ const fetchPaginationData = async (searchParams: Record<string, string | string[
   // Calculate skip value based on page and pageSize
   const skip = clamp(pageInt - 1, 0, 1e3) * pageSizeInt;
 
-  const [totalCount, allFunctionSubscriptions] = await Promise.all([
+  const [totalCount, allCollectionSubscriptions] = await Promise.all([
     // Fetch the total count of functions
-    db.functionSubscription.count({
+    db.collectionSubscription.count({
       where: {
         userId,
       },
     }),
 
     // Fetch all functions that the user has subscribed to
-    db.functionSubscription.findMany({
+    db.collectionSubscription.findMany({
       where: {
         userId: userId,
       },
-      include: {
-        function: {
+      select: {
+        id: true,
+        collection: {
           select: {
             id: true,
+            name: true,
             slug: true,
-            code: false,
             description: true,
-            httpVerb: true,
-            isPrivate: true,
-            isPublished: true,
-            ownerUserId: true,
-            createdAt: false,
-            updatedAt: true,
-            tags: {
-              select: {
-                id: true,
-                name: true,
-                functionId: false,
-                function: false,
-              }
-            },
             owner: {
               select: {
-                id: true,
                 profile: {
                   select: {
-                    id: true,
                     userName: true,
                   }
                 }
               }
+            },
+            functions: {
+              select: {
+                id: true,
+                slug: true,
+                code: false,
+                description: false,
+                httpVerb: false,
+                isPrivate: true,
+                isPublished: true,
+                ownerUserId: false,
+                createdAt: false,
+                updatedAt: false,
+                tags: {
+                  select: {
+                    id: true,
+                    name: true,
+                    functionId: false,
+                    function: false,
+                  }
+                },
+                arguments: false
+              },
             }
-          }
-        }
+          },
+        },
       },
       skip,
       take: pageSizeInt,
@@ -78,8 +86,8 @@ const fetchPaginationData = async (searchParams: Record<string, string | string[
   const hasNext = (skip + pageSizeInt) < totalRecords;
   const hasPrevious = pageInt > 1;
 
-  const result: FetchFunctionSubscriptionListPaginationType = {
-    records: allFunctionSubscriptions,
+  const result: FetchCollectionSubscriptionListPaginationType = {
+    records: allCollectionSubscriptions,
     page: pageInt,
     pageSize: pageSizeInt,
     totalRecords,
@@ -102,19 +110,16 @@ export default async function Page(
     return notFound();
   }
 
-
   const paginationData = await fetchPaginationData(searchParams, authSession.user.id),
     { records, ...paginationProps } = paginationData;
 
-  const functionListGridClasses = cn({
-    'grid gap-4 grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-6': true
-  });
+  const allCollections = records.map((record) => record.collection);
 
   return (
     <div className="block" data-testid="functions-subscriptions-page">
       <div className="sm:flex sm:items-center">
         <div className="sm:flex-auto">
-          <h1 className="text-base font-semibold leading-6">Function Subscriptions</h1>
+          <h1 className="text-base font-semibold leading-6">Collection Subscriptions</h1>
           <p className="text-xs text-gray-500 font-sans mt-1">
             # records: <strong>{paginationProps.totalRecords.toLocaleString(undefined, { minimumFractionDigits: 0 })}</strong>
           </p>
@@ -124,13 +129,7 @@ export default async function Page(
       </div>
       <div className="mt-4">
         <BakaPagination className="mb-4" {...paginationProps} />
-        <div className={functionListGridClasses}>
-          {
-            records.length > 0 && (
-              <FunctionsList functionSubscriptions={records} />
-            )
-          }
-        </div>
+        <CollectionList collections={allCollections} />
       </div>
     </div>
   );
