+++ app/(protected)/subscriptions/page.tsx
@@ -1,5 +1,5 @@
 import { RedirectType, redirect } from 'next/navigation';
 
 export default async function Page() {
-  redirect('/mine/subscriptions/collections', RedirectType.replace);
+  redirect('/subscriptions/collections', RedirectType.replace);
 }
